% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_species_data.R
\name{split_species_data}
\alias{split_species_data}
\title{Split a GBIF-like occurrence table into a list by species}
\usage{
split_species_data(occ_data)
}
\arguments{
\item{occ_data}{A \code{data.frame} with at least \code{scientificName}, \code{decimalLatitude}, \code{decimalLongitude}.}
}
\value{
A named \code{list(data.frame, ...)} with one element per species.
}
\description{
Split a GBIF-like occurrence table into a list by species
}
