% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vsc_plot_occ_heatmap.R
\name{vsc_plot_occ_heatmap}
\alias{vsc_plot_occ_heatmap}
\title{Plot heatmap of occurrences per species and year}
\usage{
vsc_plot_occ_heatmap(occ, break_by = NULL)
}
\arguments{
\item{occ}{A data.frame from \code{vsc_read_occurrences()}.}

\item{break_by}{Integer step for legend breaks (default auto).}
}
\value{
A \code{ggplot} object.
}
\description{
Plot heatmap of occurrences per species and year
}
