% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vsc_plot_raster_with_grid.R
\name{vsc_plot_raster_with_grid}
\alias{vsc_plot_raster_with_grid}
\title{Quick raster + grid plot (ggplot)}
\usage{
vsc_plot_raster_with_grid(r, grid, layer = 1)
}
\arguments{
\item{r}{A \code{SpatRaster} (single-layer) to plot as background.}

\item{grid}{An \code{sf} polygon grid.}

\item{layer}{Optional layer name or index if \code{r} has multiple layers.}
}
\value{
A \code{ggplot} object.
}
\description{
Quick raster + grid plot (ggplot)
}
