% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vsc_read_occurrences.R
\name{vsc_read_occurrences}
\alias{vsc_read_occurrences}
\title{Read GBIF occurrence file and select columns}
\usage{
vsc_read_occurrences(file, year_min = 2000, year_max = 2017)
}
\arguments{
\item{file}{Path to a GBIF TSV/TXT file with standard Darwin Core columns.}

\item{year_min, year_max}{Integer bounds to filter occurrences (inclusive).}
}
\value{
A \code{data.frame} with columns: scientificName, decimalLatitude, decimalLongitude, year.
}
\description{
Read GBIF occurrence file and select columns
}
